refactor(validators): break facade type dependency
