+++ src/core/validators/character-card.ts
@@ -1,4 +1,4 @@
-import type { ValidationResult } from "./index";
+import type { ValidationResult } from "./types";
 import {
   countListItems,
   countListItemsUnder,
